Bug 753776: Fix opening new pages in targetted frames.
Fixed the bug (753776) where newly opened windows targetted at named
frames were not opening in the correct windows.  This was fixed by the
changes to pass the default page and the changed returned page from
click().

Changed returned page from click() method to work with new targets.
That is, instead of just returning changes to the immediate window, it
returns any window that is opened.  So, if you call click() on an
element that has a target set, you will get the page in the window for
that target.  This is particularly needed for the _blank target where
you couldn't find the new window without this change since it is not
named.

To change the returned page from click() a stack of "first" windows
opened was added.  As each nested JavaScript is executed, the window is
pushed onto this stack so the corresponding window can be returned.

Also needed to support the returned page for click() was passing a
default page to doClickAction().  This affected all classes that
implement doClickAction().

Removed redundant calls to setEnclosedPage since the getPage methods
call this method.

Fixed a bug where windows that were overwritten were not cleaned up so
they would be returned from getWebWindowByName.  Added Page.cleanUp() to
remove old windows from the list of open windows so the look up of
window names won't get windows that don't exist anymore.  This method
was added to all derived classes as well.

Added WebWindow.getTopWindow() and getParentWindow().  These methods
were added to all derived classes as well.  Changed the JavaScript
Window class methods to use the WebWindow methods.

Added support for the remaining special target names (_blank, _parent,
_top in addition to _self that was already supported).  These special
target names are supported for a, area and form elements as well as the
JavaScript open() function.  Factored out new method, openTargetWindow
to encapsulate handling of window targets.  Added 2 overloaded getPage
methods that take target names.  Called these new getPage methods for a,
area, and form elements so all these elements support the special
targets.  Used the factored out method in openWindow used for JavaScript
open() so it supports all special targets.

Added support for DOM API node navigation methods in HtmlElement so they
are available within test cases.  This includes getParentNode,
getNextSibling, getPreviousSibling, getFirstChild, getLastChild,
getNodeType, getNodeName, and getNodeValue.  Changed the JavaScript
HTMLElement class so the corresponding methods use the HtmlElement
methods.  Had to add a workaround for Document.jsGet_parentNode to
return null due to HtmlPage being both the HTML and Document nodes in
this implementation.

Fixed a bug in HtmlPage.initializeHtmlElementCreatorsIfNeeded where a
base element without an href attribute was causing a
MalformedURLException to be thrown.  This can be caused when the base
element just sets a default target.  Similarly check for empty href
attribute in a and area elements.  (The HTML 4 standard does not allow
the action attribute to be optional in form elements.)

Added support for base elements to set a default target using the target
attribute.  This affects a, area and form elements.  This is implemented
in getResolvedTarget().

Added new tests and extended existing tests to support the new features.
A couple tests were changed due to the difference in the returned page
from click().

I couldn't get maven htmlunit:checkstyle to work.  It kept producing
this error despite my attempts to work around it.  When I removed the
optional dependency it said checkstyle was not defined.

Attempting to download js-1.5R4.1.jar.
WARNING: Failed to download js-1.5R4.1.jar.
The build cannot continue because of the following unsatisfied
dependency:

js-1.5R4.1.jar (try downloading from http://www.mozilla.org/rhino)


git-svn-id: svn://svn.code.sf.net/p/htmlunit/code/trunk@428 5f5364db-9458-4db8-a492-e30667be6df6