Added support for JavaScript editing of select options: add new options, replace options, remove options, shorten options list, set value, and set text.  These features also available through HtmlOption.
Fixed support for option label attribute so label value is returned as
text if label is set.

The editing of select options was implemented by adding an OptionsArray
class that provides the special semantics.  Also added methods to allow
attributes to be set.

One complication with adding options is the Option object is newed from
within a script, which means it isn't properly initialized with the page
info and DOM object.  The only way I found to handle this is to set
these when the Option object is added.  This also required the
constructor parameters to be cached and set in the DOM object when it is
set.


git-svn-id: svn://svn.code.sf.net/p/htmlunit/code/trunk@451 5f5364db-9458-4db8-a492-e30667be6df6